Removed empty strings and modularized the code
