+++ detailViewUtils.js
@@ -1,0 +1,7 @@
+import { detailscontent } from './data.js';
+
+export const openDetailView = (index) => {
+  const card = document.getElementById(`card-${index}`);
+  const detailViewImage = detailscontent[index].detailimage;
+  card.style.backgroundImage = `url('${detailViewImage}')`;
+};
